Add files via upload
This is a working project based on C#.net and SQL Server. Works using the hardware interface device and the coding of hardware interface is already provided. The guidelines are uploaded via the PDF. The above are the configuration files needed for the hardware interface. The coding is already done, You just need to set up the connections and instructions are given in  the PDF.